add clustering methods
